requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
mysqlclient==2.2.4
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
